+++ Tests/Unit/MSFT_xSQLServerAlwaysOnAvailabilityGroup.Tests.ps1
@@ -19,8 +19,6 @@ $TestEnvironment = Initialize-TestEnvironment `
 Import-Module -Name ( Join-Path -Path ( Join-Path -Path $PSScriptRoot -ChildPath Stubs ) -ChildPath SQLPSStub.psm1 ) -Force -Global
 Add-Type -Path ( Join-Path -Path ( Join-Path -Path $PSScriptRoot -ChildPath Stubs ) -ChildPath SMO.cs )
 
-
-
 # Begin Testing
 try
 {
@@ -34,6 +32,8 @@ try
             AvailabilityMode = 'AsynchronousCommit'
             BackupPriority = 50
             BasicAvailabilityGroup = $false
+            DatabaseHealthTrigger = $true
+            DtcSupportEnabled = $true
             ConnectionModeInPrimaryRole = 'AllowAllConnections'
             ConnectionModeInSecondaryRole = 'AllowNoConnections'
             FailureConditionLevel = 'OnServerDown'
@@ -51,6 +51,8 @@ try
             AvailabilityMode = 'AsynchronousCommit'
             BackupPriority = 50
             BasicAvailabilityGroup = $false
+            DatabaseHealthTrigger = $true
+            DtcSupportEnabled = $true
             ConnectionModeInPrimaryRole = 'AllowAllConnections'
             ConnectionModeInSecondaryRole = 'AllowNoConnections'
             FailureConditionLevel = 'OnServerDown'
@@ -109,6 +111,17 @@ try
             $mockAvailabilityGroup.LocalReplicaRole = 'Primary'
             $mockAvailabilityGroup.AvailabilityReplicas = $mockAvailabilityReplicaCollection
 
+            If ( $sqlVersion -eq 13 )
+            {
+                ForEach ( $member in ( 'BasicAvailabilityGroup','DatabaseHealthTrigger','DtcSupportEnabled' ) )
+                {
+                    If ( -not ( $mockAvailabilityGroup | get-member -Name $member ) )
+                    {
+                        $mockAvailabilityGroup | Add-Member -NotePropertyName $member -NotePropertyValue $False
+                    }
+                }
+            }
+
             # Create the availability group collectionto store the availability groups in
             $mockAvailabilityGroupCollection = New-Object Microsoft.SqlServer.Management.Smo.AvailabilityGroupCollection
             $mockAvailabilityGroupCollection.Add($mockAvailabilityGroup)
@@ -146,6 +159,7 @@ try
             $mockServerObject.Name = 'Server1'
             $mockServerObject.NetName = 'Server1'
             $mockServerObject.Roles = @{}
+            $mockServerObject.EngineEdition = 'EnterpriseOrDeveloper'
             $mockServerObject.Version = @{
                 Major = $sqlVersion
             }
@@ -426,6 +440,8 @@ try
                     $result.FailoverMode | Should Be 'Manual'
                     $result.HealthCheckTimeout | Should Be 30000
                     $result.BasicAvailabilityGroup | Should BeNullOrEmpty
+                    $result.DatabaseHealthTrigger | Should BeNullOrEmpty
+                    $result.DtcSupportEnabled | Should BeNullOrEmpty
 
                     Assert-MockCalled -CommandName Connect-SQL -Scope It -Times 1 -Exactly
                 }
@@ -456,6 +472,8 @@ try
                     $result.FailoverMode | Should Be 'Manual'
                     $result.HealthCheckTimeout | Should Be 30000
                     $result.BasicAvailabilityGroup | Should BeNullOrEmpty
+                    $result.DatabaseHealthTrigger | Should BeNullOrEmpty
+                    $result.DtcSupportEnabled | Should BeNullOrEmpty
 
                     Assert-MockCalled -CommandName Connect-SQL -Scope It -Times 1 -Exactly
                 }
@@ -486,6 +504,8 @@ try
                     $result.FailoverMode | Should Be 'Manual'
                     $result.HealthCheckTimeout | Should Be 30000
                     $result.BasicAvailabilityGroup | Should Be $false
+                    $result.DatabaseHealthTrigger | Should Be $false
+                    $result.DtcSupportEnabled | Should Be $false
 
                     Assert-MockCalled -CommandName Connect-SQL -Scope It -Times 1 -Exactly
                 }
@@ -516,6 +536,8 @@ try
                     $result.FailoverMode | Should Be 'Manual'
                     $result.HealthCheckTimeout | Should Be 30000
                     $result.BasicAvailabilityGroup | Should Be $false
+                    $result.DatabaseHealthTrigger | Should Be $false
+                    $result.DtcSupportEnabled | Should Be $false
 
                     Assert-MockCalled -CommandName Connect-SQL -Scope It -Times 1 -Exactly
                 }